Integrates new Testcases into AASServerComponentTest
Description:
- Moves tests to AASServerComponentTest
- Modifies AASServerComponentTest to adapt new Test cases
- Refactors AASServerComponentTest
- Deletes TestRegAndDeregOfAASAndSMFromDBOnComponentStartAndStop

Moves tests to AASServerComponentTest
